# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Color Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}": {
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Delete Session",
        "operationId": "delete_session_v1_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/photos": {
      "post": {
        "tags": [
          "photos"
        ],
        "summary": "Register Photo",
        "operationId": "register_photo_v1_sessions__session_id__photos_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/analyze": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze",
        "operationId": "analyze_v1_sessions__session_id__analyze_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/status": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Status",
        "operationId": "status_v1_sessions__session_id__status_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sessions/{session_id}/result": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Result",
        "operationId": "result_v1_sessions__session_id__result_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "400": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error response with stable error_code for client handling.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisResult": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "top_2_seasons": {
            "prefixItems": [
              {
                "type": "string",
                "enum": [
                  "Spring",
                  "Summer",
                  "Autumn",
                  "Winter"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "Spring",
                  "Summer",
                  "Autumn",
                  "Winter"
                ]
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Top 2 Seasons"
          },
          "scorecard": {
            "$ref": "#/components/schemas/Scorecard"
          },
          "reliability": {
            "$ref": "#/components/schemas/Reliability"
          },
          "result_state": {
            "type": "string",
            "enum": [
              "ok",
              "ok_low_reliability",
              "insufficient_photos",
              "no_face_detected",
              "multiple_subjects",
              "filter_suspected",
              "failed"
            ],
            "title": "Result State"
          },
          "trace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trace"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "top_2_seasons",
          "scorecard",
          "reliability",
          "result_state",
          "trace"
        ],
        "title": "AnalysisResult"
      },
      "AnalyzeRequest": {
        "properties": {
          "force_recompute": {
            "type": "boolean",
            "title": "Force Recompute",
            "default": false
          }
        },
        "type": "object",
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "AnalyzeResponse"
      },
      "PhotoRegisterRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 25000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "mime_type",
          "size_bytes"
        ],
        "title": "PhotoRegisterRequest"
      },
      "PhotoRegisterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "session_not_pending",
                "photo_limit_exceeded",
                "invalid_filename",
                "unsupported_mime_type"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "upload_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Url"
          },
          "upload_fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "accepted",
          "reasons"
        ],
        "title": "PhotoRegisterResponse"
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "error_code": {
            "type": "string",
            "enum": [
              "invalid_session",
              "session_not_found",
              "session_deleted",
              "insufficient_photos",
              "already_running",
              "already_complete",
              "result_not_ready",
              "forbidden",
              "invalid_request",
              "internal_error",
              "rate_limit_exceeded"
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "error_code"
        ],
        "title": "ProblemDetail"
      },
      "Reliability": {
        "properties": {
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "bucket": {
            "type": "string",
            "enum": [
              "High",
              "Medium",
              "Low"
            ],
            "title": "Bucket"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "score",
          "bucket",
          "reasons"
        ],
        "title": "Reliability"
      },
      "Scorecard": {
        "properties": {
          "warmth": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Warmth"
          },
          "value": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Value"
          },
          "chroma": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Chroma"
          },
          "contrast": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Contrast"
          }
        },
        "type": "object",
        "required": [
          "warmth",
          "value",
          "chroma",
          "contrast"
        ],
        "title": "Scorecard"
      },
...
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "complete",
              "failed",
              "deleted"
            ],
            "title": "Status"
          },
          "result_state": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ok",
                  "ok_low_reliability",
                  "insufficient_photos",
                  "no_face_detected",
                  "multiple_subjects",
                  "filter_suspected",
                  "failed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Result State"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "StatusResponse"
      }
    }
  }
}